requirements.txt
aiohttp==0.16.5
alabaster==0.7.5
Babel==1.3
chardet==2.3.0
docutils==0.12
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
pytz==2015.4
requests==2.7.0
RPi.GPIO==0.5.11
Rx==1.2.1
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
wheel==0.24.0
